tests

+++ htdocs/core/lib/price.lib.php
@@ -273,15 +273,15 @@ function calcul_price_total($qty, $pu, $remise_percent_ligne, $txtva, $uselocalt
 
 	if ($uselocaltax1_rate && $apply_tax) {
 		$tmpResult14 = numTruncate(price2num(($tot_sans_remise_withvat * (1 + ($localtax1_rate / 100))), ''), $dolMaxDecimalUnits);
-		$result[14] = price2num($tmpResult14 - $tot_sans_remise_withvat, 'MT');
+		$result[14] = (float) price2num($tmpResult14 - $tot_sans_remise_withvat, 'MT');
 		$localtaxes[0] += $result[14];
 
 		$tmpResult9 = numTruncate(price2num(($tot_avec_remise_withvat * (1 + ($localtax1_rate / 100))), ''), $dolMaxDecimalUnits);
-		$result[9] = price2num($tmpResult9 - $tot_avec_remise_withvat, 'MT');
+		$result[9] = (float) price2num($tmpResult9 - $tot_avec_remise_withvat, 'MT');
 		$localtaxes[1] += $result[9];
 
 		$tmpResult11 = numTruncate(price2num(($pu_withmainvat * (1 + ($localtax1_rate / 100))), ''), $dolMaxDecimalUnits);
-		$result[11] = price2num($tmpResult11 - $pu_withmainvat, 'MU');
+		$result[11] = (float) price2num($tmpResult11 - $pu_withmainvat, 'MU');
 		$localtaxes[2] += $result[11];
 	}
 
@@ -303,15 +303,15 @@ function calcul_price_total($qty, $pu, $remise_percent_ligne, $txtva, $uselocalt
 	}
 	if ($uselocaltax2_rate && $apply_tax) {
 		$tmpResul15 = numTruncate(price2num(($tot_sans_remise_withvat * (1 + ($localtax2_rate / 100))), ''), $dolMaxDecimalUnits);
-		$result[15] = price2num($tmpResul15 - $tot_sans_remise_withvat, 'MT');
+		$result[15] = (float) price2num($tmpResul15 - $tot_sans_remise_withvat, 'MT');
 		$localtaxes[0] += $result[15];
 
 		$tmpResult10 = numTruncate(price2num(($tot_avec_remise_withvat * (1 + ($localtax2_rate / 100))), ''), $dolMaxDecimalUnits);
-		$result[10] = price2num($tmpResult10 - $tot_avec_remise_withvat, 'MT');
+		$result[10] = (float) price2num($tmpResult10 - $tot_avec_remise_withvat, 'MT');
 		$localtaxes[1] += $result[10];
 
 		$tmpResult12 = numTruncate(price2num(($pu_withmainvat * (1 + ($localtax2_rate / 100))), ''), $dolMaxDecimalUnits);
-		$result[12] = price2num($tmpResult12 - $pu_withmainvat, 'MU');
+		$result[12] = (float) price2num($tmpResult12 - $pu_withmainvat, 'MU');
 		$localtaxes[2] += $result[12];
 	}
 
@@ -322,23 +322,23 @@ function calcul_price_total($qty, $pu, $remise_percent_ligne, $txtva, $uselocalt
 		// Calculer le montant total ttc implique de faire pu_ttc * qte et l'arrondir
 		// Faire pu * qte * (1+pct_taxe) créé une erreur d'accumulation de la plus petite décimale
 		// en ajoutant des quantités.
-		$result[6] = price2num($tot_sans_remise, 'MT');
+		$result[6] = (float) price2num($tot_sans_remise, 'MT');
 
 		$tmpResult8 = numTruncate(price2num($tot_sans_remise * (1 + ((($info_bits & 1) ? 0 : $txtva) / 100)), ''), $dolMaxDecimalUnits);
 		$tmpREsult8bis = numTruncate(price2num($tot_sans_remise * (1 + ($txtva / 100)), ''), $dolMaxDecimalUnits);
 
-		$result[8] = price2num($tmpResult8 + $localtaxes[0], 'MT'); // Selon TVA NPR ou non
-		$result8bis = price2num($tmpREsult8bis + $localtaxes[0], 'MT'); // Si TVA consideree normal (non NPR)
-		$result[7] = price2num((float) $result8bis - ((float) $result[6] + $localtaxes[0]), 'MT');
+		$result[8] = (float) price2num($tmpResult8 + $localtaxes[0], 'MT'); // Selon TVA NPR ou non
+		$result8bis = (float) price2num($tmpREsult8bis + $localtaxes[0], 'MT'); // Si TVA consideree normal (non NPR)
+		$result[7] = (float) price2num((float) $result8bis - ((float) $result[6] + $localtaxes[0]), 'MT');
 
-		$result[0] = price2num($tot_avec_remise, 'MT');
+		$result[0] = (float) price2num($tot_avec_remise, 'MT');
 		// Le terme $tot_avec_remise * (1 + ((($info_bits & 1) ? 0 : $txtva) / 100) doit être tronqué au
 		// nombre de décimales des totaux
-		$tmpResult2 = numTruncate(price2num($tot_avec_remise * (1 + ((($info_bits & 1) ? 0 : $txtva) / 100)), ''), $dolMaxDecimalUnits);
-		$tmpResult2bis = numTruncate(price2num($tot_avec_remise * (1 + ($txtva / 100)), ''), $dolMaxDecimalUnits);
+		$tmpResult2 = (float) numTruncate(price2num($tot_avec_remise * (1 + ((($info_bits & 1) ? 0 : $txtva) / 100)), ''), $dolMaxDecimalUnits);
+		$tmpResult2bis = (float) numTruncate(price2num($tot_avec_remise * (1 + ($txtva / 100)), ''), $dolMaxDecimalUnits);
 
-		$result[2] = price2num($tmpResult2 + $localtaxes[1], 'MT'); // Selon TVA NPR ou non
-		$result2bis = price2num($tmpResult2bis + $localtaxes[1], 'MT'); // Si TVA consideree normal (non NPR)
+		$result[2] = (float) price2num($tmpResult2 + $localtaxes[1], 'MT'); // Selon TVA NPR ou non
+		$result2bis = (float) price2num($tmpResult2bis + $localtaxes[1], 'MT'); // Si TVA consideree normal (non NPR)
 		$result[1] = price2num((float) $result2bis - ((float) $result[0] + $localtaxes[1]), 'MT'); // Total VAT = TTC - (HT + localtax)
 
 		$result[3] = price2num($pu, 'MU');
